Добавит изменения в main.js
